requirements.txt
asgiref==3.3.1
Django==3.1.4
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-google-translate==1.1
django-js-asset==1.2.2
pytz==2020.4
sqlparse==0.4.1
